combined all the processing code in preprocessing.py.
