+++ .gitignore
@@ -1724,3 +1724,4 @@ data/Therapeutic_Procedures/Procedure_Files/Raktamokshan`a_files/header.htm
 data/Therapeutic_Procedures/Procedure_Files/Samsarjana Krama_files/header.htm
 data/Therapeutic_Procedures/Procedure_Files/Vran`a Lekhana_files/header.htm
 .gitignore
+how2convert_clean_html2table_and_text.md
