--- processed_data/data/Articles/Chikitsaa/Aamadosha.htm
@@ -1,71 +1,0 @@
-<html xmlns:v="urn:schemas-microsoft-com:vml">
-
-<body>
-<div id="NavBar">
-</div>
-<div>
-<p>Aamadosha Chikitsaa</p>
-<p>Reference: </p>
-<p>C.Vi.2,
-A.H.Su.13</p>
-<p></p>
-<p>Introduction: </p>
-<p>Aama is an undigested
-food. Due to lack of Agnibala (Weak
-digestive power), the partially undigested food, which remains in Aantra, vitiated due to Dosha is called as Aama. It is the root cause for the development of majority of
-intrinsic diseases.</p>
-<p></p>
-<p>Saamaanya Chikitsaa:</p>
-<p>Apatarpan`a is advisable as primary treatment.</p>
-<p>Apatarpan`a is not advisable always. Following
-regime should be administered.</p>
-<p>Paachana, Deepana, Snehana and
-Svedana followed by the appropriate type of S`hodhana should be administered. For Deepana Paachana Sneha Dravya having Paachana property predominantly and Deepana property partially. </p>
-<p>Paachana results in digestion of Aama. Due to Deepana Aaama gets separated from the Dhaatu. Snehana results in Utkleshana.
-Svedana administered after this leads the movement of Aama towards the Kosht`ha.</p>
-<p>This Aama should now be eliminated out of the
-body through the nearest possible path by administering appropriate S`hodhana.</p>
-<p>S`hodhana </p>
-<p>For Aamaas`hayatha Aama -Through oral root</p>
-<p>For Urdhvajatrugata Aama - Through nasal
-root.</p>
-<p>For Pakvaashayastha Aama - Through anal
-root.</p>
-<p></p>
-<p>Aavasthika Chikitsaa: </p>
-<p></p>
-<p></p>
-
-<p></p>
-<p>Special Precautions (Chikitsaa
-Nishedha): </p>
-<p>S`hodhana Chikitsaa should be avoided in following
-conditions.</p>
-<p>S`hodhana
-Chikitsaa should be
-avoided in Saamadosha which are scattered
-all over the body (Leena) or clung to
-Dhaatu or which are not mobile (Anutklishat`a).</p>
-<p>Saamadosha which have spontaneously become mobile
-either upward or downwards should not be treated with S`hamana Chikitsaa. (Saama
-Dosha which have become spontaneously mobile and being eliminated from
-oral or anus should not be arrested
-in the initial stage)</p>
-<p></p>
-<p></p>
-<p>Properties of usual diet:</p>
-<p>Normally diet should be
-digested properly and there should not be any symptom like Hridaya Uparodha, Kukshee
-Peed`ana, Paars`hva Vipaat`ana, Udara Gaurava, or Indriya Apreen`ana. If diet is taken in a small amount, which does
-not fulfill hunger, then it is called Heenamaatraas`hana.
-</p>
-<p></p>
-<p>Search Anveshaka:</p>
-<p>Search Kalpa, Dravya, Aahaara, Vihaara, Chikitsaa Karma </p>
-<p>Vyaadhi = Aamadosha</p>
-<p>Karma = Paachana, Deepana</p>
-<p></p>
-<p></p>
-</div>
-</body>
-</html>
