+++ processed_data/data/Articles/Nidaana/Aamadosha.htm
@@ -1,0 +1,113 @@
+<html xmlns="http://www.w3.org/TR/REC-html40" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="urn:schemas-microsoft-com:office:word">
+<body>
+<div id="NavBar">
+</div>
+<div>
+<p>Aamadosha</p>
+<p></p>
+<p>References:
+</p>
+<p>C.Chi.15, C.Vi.2, A.H.Su.13</p>
+<p> </p>
+<p>Introduction:</p>
+<p>Heavy diet or food in
+large quantity hampers Agni (digestive
+capacity). That leads to formation of Aama,
+vitiation of Dosha and manifestation
+of symptoms, collectively called as Aamadosha.
+</p>
+<p></p>
+<p>Definition of Aama: </p>
+<p>Ushman`o Alpa Balatvena Dhaatum Aadyam
+Apaachitam |</p>
+<p>Dusht`am Aamaas`hayagatam Rasam Aamam
+Prachakshate || A. H.
+Su. 13/25</p>
+<p></p>
+<p>Aama
+is formed due to decreased digestive capacity. In the process of digestion, Rasa
+Dhaatu remains
+undigested and gets vitiated. This vitiated Rasa remains in the Aamaas`haya
+causing various symptoms and diseases. This is called as Aama Dosha. </p>
+<p></p>
+<p></p>
+<p>Causes of Aamadosha:</p>
+<p>Diet consumed in
+large amount than the hunger, it is called Atimaatraas`hana, which is cause of
+aggravation of all the three Dosha. </p>
+<p>When a person
+consumes food in a large amount, and drinks water after consumption, due to the
+irritation caused by the excess food all the three Dosha get aggravated. </p>
+<p>Not only excess food
+quantity is responsible for Aama Dosha
+but the food having attributes like Guru,
+Rooksha, S`heeta etc. also causes Aama
+Dosha. Though the food is taken in proper amount, factors like Chintaa etc. are also responsible for Aama Dosha </p>
+<p>If all the Dosha are vitiated and they get accumulated,
+then intermingle to produce Aama as
+poison is formed from the Kodrava. </p>
+<p></p>
+<p>Sampraapti of Aamadosha:</p>
+<p>Agni
+Dusht`ee is basic
+cause of Aamadosha. </p>
+<p>Due to hampered digestive capacity, though light
+food is consumed it is not digested then it resembles to S`hukta (fermented substance) and turns to Anna Visha. </p>
+<p>These Dosha
+get mixed with the undigested food. They remain there without any movement (Alasaka) or suddenly they get purged out
+either through the mouth or through the anus (Visoochikaa). Thus, they create different types of Vyaadhi. </p>
+<p>Anna Visha (food turned into poison) gets associated with Pitta, then it manifests with Daaha, Trishn`aa, Mukhaamaya (diseases
+of oral cavity), Amlapitta and other
+diseases due to Pitta. If it gets
+associated with Kapha, then it
+manifests in Yakshmaa, Peenasa, Meha and other Kapha predominant diseases. </p>
+<p><u></u></p>
+<p></p>
+<p></p>
+<p></p>
+<p>General Properties of Aama: </p>
+<ol start="1" type="1">
+<li>Drava (liquid)</li>
+<li>Guru (heavy to digest)</li>
+<li>Aneka
+     Varn`a (having different
+     colours)</li>
+<li>Snigdha (unctuous)</li>
+<li>Pichchhila (slimmy)</li>
+<li>Tantumad
+     </li>
+<li>Anubaddha
+     S`hoola (pain)</li>
+<li>Durgandhee (with bad odour)</li>
+</ol>
+<p></p>
+<p>Clinical presentation of Aamadosha</p>
+<p>When the diseases are
+either caused by Saama Dosha, Saama
+Dooshya / Dhaatu or Saama Mala or
+have stage of Aama association, are
+called as Saama. </p>
+<p>Aama Dosha
+can manifests with two types Visoochikaa
+and Alasaka.</p>
+<p>Aamadosha may present following symptoms:</p>
+<p>Visht`ambha</p>
+<p>Sadana</p>
+<p>S`hiroruk</p>
+<p>Moorchchhaa</p>
+<p>Bhrama</p>
+<p>Prisht`ha Graha</p>
+<p>Kat`i Graha</p>
+<p>Jrimbhaa</p>
+<p>Angamarda</p>
+<p>Trishn`aa</p>
+<p>Jvara</p>
+<p>Chhardi</p>
+<p>Pravaahan`a</p>
+<p>Arochaka </p>
+<p>Avipaaka</p>
+<p></p>
+<p></p>
+</div>
+</body>
+</html>
