+++ processed_data/data/Articles/Nidaana/Aamadosha.htm.txt
@@ -1,0 +1,113 @@
+
+
+
+
+
+Aamadosha
+
+References:
+
+C.Chi.15, C.Vi.2, A.H.Su.13
+ 
+Introduction:
+Heavy diet or food in
+large quantity hampers Agni (digestive
+capacity). That leads to formation of Aama,
+vitiation of Dosha and manifestation
+of symptoms, collectively called as Aamadosha.
+
+
+Definition of Aama: 
+Ushman`o Alpa Balatvena Dhaatum Aadyam
+Apaachitam |
+Dusht`am Aamaas`hayagatam Rasam Aamam
+Prachakshate || A. H.
+Su. 13/25
+
+Aama
+is formed due to decreased digestive capacity. In the process of digestion, Rasa
+Dhaatu remains
+undigested and gets vitiated. This vitiated Rasa remains in the Aamaas`haya
+causing various symptoms and diseases. This is called as Aama Dosha. 
+
+
+Causes of Aamadosha:
+Diet consumed in
+large amount than the hunger, it is called Atimaatraas`hana, which is cause of
+aggravation of all the three Dosha. 
+When a person
+consumes food in a large amount, and drinks water after consumption, due to the
+irritation caused by the excess food all the three Dosha get aggravated. 
+Not only excess food
+quantity is responsible for Aama Dosha
+but the food having attributes like Guru,
+Rooksha, S`heeta etc. also causes Aama
+Dosha. Though the food is taken in proper amount, factors like Chintaa etc. are also responsible for Aama Dosha 
+If all the Dosha are vitiated and they get accumulated,
+then intermingle to produce Aama as
+poison is formed from the Kodrava. 
+
+Sampraapti of Aamadosha:
+Agni
+Dusht`ee is basic
+cause of Aamadosha. 
+Due to hampered digestive capacity, though light
+food is consumed it is not digested then it resembles to S`hukta (fermented substance) and turns to Anna Visha. 
+These Dosha
+get mixed with the undigested food. They remain there without any movement (Alasaka) or suddenly they get purged out
+either through the mouth or through the anus (Visoochikaa). Thus, they create different types of Vyaadhi. 
+Anna Visha (food turned into poison) gets associated with Pitta, then it manifests with Daaha, Trishn`aa, Mukhaamaya (diseases
+of oral cavity), Amlapitta and other
+diseases due to Pitta. If it gets
+associated with Kapha, then it
+manifests in Yakshmaa, Peenasa, Meha and other Kapha predominant diseases. 
+
+
+
+
+General Properties of Aama: 
+
+Drava (liquid)
+Guru (heavy to digest)
+Aneka
+     Varn`a (having different
+     colours)
+Snigdha (unctuous)
+Pichchhila (slimmy)
+Tantumad
+     
+Anubaddha
+     S`hoola (pain)
+Durgandhee (with bad odour)
+
+
+Clinical presentation of Aamadosha
+When the diseases are
+either caused by Saama Dosha, Saama
+Dooshya / Dhaatu or Saama Mala or
+have stage of Aama association, are
+called as Saama. 
+Aama Dosha
+can manifests with two types Visoochikaa
+and Alasaka.
+Aamadosha may present following symptoms:
+Visht`ambha
+Sadana
+S`hiroruk
+Moorchchhaa
+Bhrama
+Prisht`ha Graha
+Kat`i Graha
+Jrimbhaa
+Angamarda
+Trishn`aa
+Jvara
+Chhardi
+Pravaahan`a
+Arochaka 
+Avipaaka
+
+
+
+
+
