+++ processed_data/data/Articles/Other/Mantra.htm
@@ -1,0 +1,37 @@
+<html xmlns="http://www.w3.org/TR/REC-html40" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="urn:schemas-microsoft-com:office:word">
+<body>
+<div id="NavBar">
+</div>
+<h3>Chikitsaa Prakaara: Mantra</h3>
+<p></p>
+<p>Prakaara
+Paribhaashaa:</p>
+<p>+I…Æº…®…⁄½˝&amp;
+ª…ºª……‰c……Æ‰h…
+¥ª……Êv…¯þ{…∂……®ª…Êi… n‰¸¥……n¸ª…õ…
+|…º…z…… ¶…¥…Êi… * Ayurved S`habdakos`ha </p>
+<p>It is a procedure in which
+chanting of special group of words or sounds by a specialist in the field of Mantra Chikitsaa are done as an
+antidote for the poison.</p>
+<p></p>
+<p>Chikitsaa
+Karma (Benefits): </p>
+<p>Create a hope in the mind of stricken patient and hope of recovery surely
+aids in control of poisonous effect so Mantra
+is excellent as Aas`hvaasana Chkitsaa
+(reassurance) for patient as well as for the near and dear relatives of the
+patient.</p>
+<p>The specialists are capable of doing Arisht`abandhana
+only by chanting Mantra.</p>
+<p>Visha Pratibandha (counteracts poison).</p>
+<p>Visha Prasara Nirodha (prevents circulation of
+poison in whole body).</p>
+<p>Other Description:</p>
+<p>Now days because of
+non availability of specialists this procedure is not in practice.</p>
+<p>One should always be
+aware of the pseudo specialists who are not masters of their art, and who have
+studied Mantra without observing the
+protocol of study or who chant imperfect word groups or use wrong notes.</p>
+</body>
+</html>
