+++ processed_data/data/Articles/Other/Mantra.htm.txt
@@ -1,0 +1,37 @@
+
+
+
+
+Chikitsaa Prakaara: Mantra
+
+Prakaara
+Paribhaashaa:
++I…Æº…®…⁄½˝&
+ª…ºª……‰c……Æ‰h…
+¥ª……Êv…¯þ{…∂……®ª…Êi… n‰¸¥……n¸ª…õ…
+|…º…z…… ¶…¥…Êi… * Ayurved S`habdakos`ha 
+It is a procedure in which
+chanting of special group of words or sounds by a specialist in the field of Mantra Chikitsaa are done as an
+antidote for the poison.
+
+Chikitsaa
+Karma (Benefits): 
+Create a hope in the mind of stricken patient and hope of recovery surely
+aids in control of poisonous effect so Mantra
+is excellent as Aas`hvaasana Chkitsaa
+(reassurance) for patient as well as for the near and dear relatives of the
+patient.
+The specialists are capable of doing Arisht`abandhana
+only by chanting Mantra.
+Visha Pratibandha (counteracts poison).
+Visha Prasara Nirodha (prevents circulation of
+poison in whole body).
+Other Description:
+Now days because of
+non availability of specialists this procedure is not in practice.
+One should always be
+aware of the pseudo specialists who are not masters of their art, and who have
+studied Mantra without observing the
+protocol of study or who chant imperfect word groups or use wrong notes.
+
+
