+++ processed_data/data/Articles/Research/Prakriti_Article.htm.txt
@@ -1,0 +1,81 @@
+
+
+
+
+
+
+Scope & Research Potential of Prakriti
+
+
+In the current era there
+is more stress on preventive health care, not only of the physical being but
+also of mental and social well being which has become the principal issue. Aayurveda is not only limited to body or
+physical symptoms but also gives a comprehensive knowledge about spiritual,
+mental and social health. Knowing your constitution thus allows you to
+appreciate the workings of your mind and body better, there by allowing in managing
+and systematizing over the traits through planned and adequate changes
+incorporated into your lifestyle. Ayurveda
+places importance on examining the Prakriti
+or the natural state of an individual first, the disease state is examined
+later. While the medicine experts analyze the disease, the Ayurvedic expert is interested in analyzing not only the disease
+but the individual who is suffering from the disease.
+
+The word Prakriti (constitution) not only gives
+perception of the anatomical features but also provides an insight into the
+deeper internal mechanism of an individual in terms of physiology, psychology
+and quality and excellence of tissues. Prakriti
+is the Ayurvedic profile or unique
+psychosomatic temperament of an individual, encompassing his or her physical,
+functional and behavioral characteristics. The varying proportions of Kapha, Pitta or Vaata Dosha determine the S`haareera Prakriti or the Dosha
+Prakriti i.e., the body type which play a pivotal role. While all
+individuals have mixed sum of the three Gun`a,
+the predominant Gun`a determines an
+individual's Maanasa Prakriti.
+Evaluation of Maanasa Prakriti, i.e.,
+psycho-spiritual profile is made by identifying the predominance of Psychic
+factors i.e., Sattva, Raja and Tama. Dhaatu are termed as body constituents. These are seven in
+number and their qualitative assessment can be performed by Saarataa Pareekshan`a. Saara is the essence of that particular constituent.
+It reflects on various body parts in positive manner by which it is assessed. 
+
+Like genetic coding, every
+individual has a different combination and is therefore a distinct entity in
+itself. That is the reason why two people react differently when exposed to a
+similar situation. Thus according to Aayurveda
+everyone is unique and each one of them has a distinct body constitution, which
+is completely different from each other. Nature of emotions, thought, and
+behavior patterns are unique to a person and are exhibited in a wide range of
+social and personal environment. Interpretation of the same is necessary for
+their behavior in cultural, social and home environment.
+
+
+Need for understanding the concept of Prakriti:
+
+By knowing our
+constitution it is possible to locate our drawbacks and live at peace with
+nature's wishes. There are some simple rules that help in the long run. The Dosha that predominate due to our
+genetic inheritance dictates certain demands in terms of diet, life style and
+environment. If we recognize this and try to follow it, we will be able to
+achieve health. 
+These are the following
+important reasons why knowing your Prakriti
+is the first step toward perfect health:-
+By understanding Prakriti and their specific strengths
+and weaknesses, you can begin to take preventive steps long before overt
+illness appears. 
+Prakriti
+makes prevention more precise. If you try to prevent every disease without
+knowing your particular predisposition, its like aiming in the dark.
+Prakriti
+makes treatment more accurate once a disease appears. 
+Study of Prakriti offers reference points for
+managing treatment decisions specific to each case. 
+Study of Prakriti helps in understanding
+variations in disease manifestation
+
+
+Scope of Prakriti:
+
+ 
+
+
+
